pyproject.toml
[project]
name = "api"
requires-python = ">=3.13"
version = "0.1.0"
dependencies = [
    "django>=5.1",
    "psycopg[binary]>=3.2",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-unfold>=0.43.0",
    "turbodrf>=0.1.12",
    "python-dotenv>=1.1.1",
    "django-extensions>=4.1",
    "django-split-settings>=1.3.2",
    "django-watchfiles>=1.1.0",
    "drf-stripe-subscription>=1.2.2",
    "django-prometheus>=2.4.1",
    "django-auditlog>=3.2.1",
    "django-easy-logging>=0.70",
    "drf-api-tracking>=1.8.4",
    "django-allauth>=65.10.0",
    "dj-rest-auth @ git+https://github.com/Mte90/dj-rest-auth.git@patch-1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-factoryboy>=2.7.0",
    "pytest-cov>=6.2.1",
    "django-cors-headers>=4.7.0",
    "ruff>=0.0.280",
    "pre-commit>=3.3.3",
]

[tool.ruff]
fix = true
line-length = 88
exclude = ["api/settings", "setup.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
filterwarnings = "ignore"
addopts = "--strict-config --strict-markers --ds=api.settings"
